src/event-handler/requirements.txt
redis==7.4.0
PyYAML==6.0.3
requests==2.33.0
httpx==0.27.2
nats-py==2.14.0